deployment/requirements.txt
streamlit
pandas
seaborn
matplotlib
plotly
Pillow
feature-engine
scikit-learn==1.2.2
tensorflow==2.12.0
